requirements.txt
streamlit
PyPDF2
openai==0.28 # for pdf chatbot
openai==1.51.0 #for image genearation
pillow